% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{APO_dml}
\alias{APO_dml}
\title{This function estimates the average potential outcomes for all treatment levels.}
\usage{
APO_dml(y, m_mat, w_mat, e_mat, cf_mat, norm = 2, cl = NULL)
}
\arguments{
\item{y}{A numeric vector containing the outcome variable.}

\item{m_mat}{n x T+1 matrix with fitted outcome values.}

\item{w_mat}{Logical matrix of treatment indicators (n x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{e_mat}{n x T+1 matrix with propensity scores.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{norm}{Controls normalization of IPW weights. \code{norm = 1)}: overall normalization, \code{norm = 2)}: normalization in each
cross-fitting fold separately (default), otherwise no normalization.}

\item{cl}{If not NULL, vector with cluster variables.}
}
\value{
Returns an \code{APO_dml} object:
         \item{results}{Point estimate and standard error of APOs.}
         \item{gamma}{n x T+1 matrix containing the APO DR scores.}
         \item{y}{Vector of outcomes used in estimation.}
         \item{m_mat}{Matrix of outcome predictions used in estimation.}
         \item{w_mat}{Matrix of treatment indicators (n x T+1) used in estimation.}
         \item{e_mat}{Matrix of propensity score used in estimation.}
         \item{cf_mat}{Matrix with k columns of indicators representing the different folds used in estimation.}
         \item{cl}{Cluster variable if specified.}
}
\description{
This function estimates the average potential outcomes for all treatment levels.
}
