% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{APO_dml_atet}
\alias{APO_dml_atet}
\title{This function calculates the average potential outcomes for the average treatment effect on the treated.}
\usage{
APO_dml_atet(y, m_mat, w_mat, e_mat, cf_mat, treated = 2, norm = 2, cl = NULL)
}
\arguments{
\item{y}{A numeric vector containing the outcome variable.}

\item{m_mat}{n x T+1 matrix with fitted outcome values.}

\item{w_mat}{Logical matrix of treatment indicators (n x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{e_mat}{n x T+1 matrix with (generalized) propensity scores.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{treated}{An integer describing the column number of the treated in w_mat for which ATET
should be calculated (default 2).}

\item{norm}{Controls normalization of IPW weights. \code{norm = 1)}: overall normalization, \code{norm = 2)}:
normalization in each cross-fitting fold separately (default), otherwise no normalization.}

\item{cl}{If not NULL, vector with cluster variables.}
}
\value{
\code{APO_dmlmt} returns a list containing the \code{results} and a matrix
         with the individual values of the doubly robust score.
}
\description{
This function calculates the average potential outcomes for the average treatment effect on the treated.
}
