% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{ATE_dml}
\alias{ATE_dml}
\title{This function calculates the average treatment effects (on the treated) for all combinations
of potentially multiple treatments.}
\usage{
ATE_dml(APO)
}
\arguments{
\item{APO}{Object of class \code{\link{APO_dml}} or \code{\link{APO_dml_atet}}.}
}
\value{
Returns an \code{ATE_dml} object:
         \item{results}{Point estimate and standard error of APOs.}
         \item{delta}{n x T+1 matrix containing the ATE DR scores.}
}
\description{
This function calculates the average treatment effects (on the treated) for all combinations
of potentially multiple treatments.
}
