% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_effects.R
\name{clan}
\alias{clan}
\title{Runs classification analysis (CLAN) for CATEs.}
\usage{
clan(cate, x, q = 5)
}
\arguments{
\item{cate}{Vector of CATEs.}

\item{x}{Covariate matrix.}

\item{q}{Number of splits (defaut 5).}
}
\value{
Matrix containing means of covariate in each group.
}
\description{
Runs classification analysis (CLAN) for CATEs.
}
