% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{CV_core}
\alias{CV_core}
\title{This function contains the core parts of the CV for Lasso and Post-Lasso}
\usage{
CV_core(x, y, w, cvgroup, list, i, lambda, ...)
}
\arguments{
\item{x}{covariate matrix to be used in CV}

\item{y}{vector of outcomes}

\item{w}{vector of weight}

\item{cvgroup}{categorical with k groups to identify folds}

\item{list}{list 1:k}

\item{i}{number of fold that is used for prediction}

\item{lambda}{series of lambdas used}

\item{...}{Pass \code{\link{glmnet}} options}
}
\value{
MSE_lasso / MSE_plasso: means squared errors for each lambda
}
\description{
This function contains the core parts of the CV for Lasso and Post-Lasso
}
\keyword{internal}
