% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_utils.R
\name{data_screen}
\alias{data_screen}
\title{Data screening}
\usage{
data_screen(data, treat = NULL, bin_cut = 0.01, corr_cut = 0.99, print = FALSE)
}
\arguments{
\item{data}{Matrix the variables to be screened.}

\item{treat}{Optional binary treatment vector if screening should be done within treatment groups}

\item{bin_cut}{Cut-off fraction under which nearly empty binary variables should be removed. Default 0.01.}

\item{corr_cut}{Cut-off above which highly correlated varialbes should be removed. Default 0.99.}

\item{print}{Shows details about the reomved variables at each step}
}
\value{
Screened matrix
}
\description{
This function takes a matrix of data and removes
1. Variables without variation
2. Dummy variables where one group is nearly empty (optional in one of both treatment groups)
3. Redundant (highly correlated variables)
}
