% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{DML_partial_linear}
\alias{DML_partial_linear}
\title{Double Machine Learning estimation of partially lienar model}
\usage{
DML_partial_linear(
  y,
  w,
  x,
  ml_w = list(create_method("forest_grf")),
  ml_y = list(create_method("forest_grf")),
  cf = 5,
  cv = 5,
  weights = FALSE,
  path = NULL,
  quiet = TRUE,
  e_hat = NULL,
  m_hat = NULL,
  cf_mat = NULL
)
}
\arguments{
\item{y}{Numeric vector containing the outcome variable.}

\item{w}{Treatment vector (binary or continuous).}

\item{x}{Covariate matrix.}

\item{ml_w}{List of methods to be used in ensemble estimation of treatment regression.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_y}{List of methods to be used in ensemble estimation of outcome regression.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{cf}{Number of cross-fitting folds for DML (default 5).}

\item{cv}{Number of cross-validation folds when estimating ensemble if more than one method is defined
in \code{ml_w} and/or \code{ml_y} (default 5).}

\item{weights}{If TRUE, prediction weights of the outcome nuisance extracted and saved (requires to provide a path).}

\item{path}{Optional path to save the \code{\link{ensemble}} objects of each cross-fit for later inspection.}

\item{quiet}{If FALSE, ensemble estimators print method that is currently running.}

\item{e_hat}{Optional vector of predicted treatment outside of the function.}

\item{m_hat}{Optional vector of predicted outcome outside of the function.}

\item{cf_mat}{Optional prespecified logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}
}
\value{
Returns a \code{DML_partial_linear} object:
         \item{results}{Point estimate, standard error, t- and p-value of estimated effect.}
         \item{e_hat}{Predicted treatment}
         \item{m_hat}{Predicted outcomes}
         \item{w}{Treatment vector used in estimation.}
         \item{y}{Vector of outcomes used in estimation.}
         \item{cf_mat}{Matrix with k columns of indicators representing the different folds used in estimation.}
         \item{path}{Path where results are stored if specified, otherwise NULL.}
}
\description{
This function estimates the parameter of interest in a partially linear model
using the residual-on-residual representation of Robinson (1988)
and flexibly estimated nuisance parameters following Chernozukov et al. (2018).
}
\references{
\itemize{
\item Robinson, P. M. (1988). Root-N-consistent semiparametric regression. Econometrica, 931-954.
\item Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen, C., Newey, W., & Robins, J. (2018).
Double/Debiased machine learning for treatment and structuralparameters.The Econometrics Journal,21(1), C1-C68
}
}
