% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_learner.R
\name{dr_core}
\alias{dr_core}
\title{Core function of \code{\link{dr_learner}}. It executes steps 2 and 4 of algorithm 1 in Knaus (2020).}
\usage{
dr_core(
  ml,
  delta,
  y,
  x,
  w_mat,
  m_mat,
  e_mat,
  cf_mat,
  xnew = NULL,
  nfolds = 5,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation.
Methods can be created by \code{\link{create_method}}.}

\item{delta}{vector of doubly robust score. E.g. create via \code{\link{ATE_dml}}.}

\item{y}{Vector of variable to be predicted.}

\item{x}{Matrix of covariates.}

\item{w_mat}{Logical matrix of treatment indicators (n x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{m_mat}{n x T+1 matrix with fitted outcome values.}

\item{e_mat}{n x T+1 matrix with propensity scores.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{xnew}{Covariate matrix of test sample}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights (default \code{nfolds=5})}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
Saved as Ensemble_Yi where i is the number of the treatment in multiple treatment settings.}

\item{quiet}{If FALSE, ensemble estimators print method that is currently running.}
}
\value{
Vector with predicted CATEs.
}
\description{
Core function of \code{\link{dr_learner}}. It executes steps 2 and 4 of algorithm 1 in Knaus (2020).
}
\references{
\itemize{
\item Knaus, M. C. (2020). Double machine learning based program evaluation under unconfoundedness.
  arXiv preprint arXiv:2003.03191.\url{http://arxiv.org/abs/2003.03191}
}
}
