% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_core}
\alias{ensemble_core}
\title{Core function of \code{\link{ensemble}}.}
\usage{
ensemble_core(ml, x_tr, y_tr, x_te, weights = FALSE, quiet = TRUE)
}
\arguments{
\item{ml}{List of methods build via \code{\link{create_method}} to be used in ensemble.}

\item{x_tr}{Covariate matrix of training sample.}

\item{y_tr}{Vector of outcomes of training sample.}

\item{x_te}{Covariate matrix of test sample.}

\item{weights}{If TRUE, weights underlying the ensemble prediction for x_te calculated.}

\item{quiet}{If FALSE, print method that is currently computed.}
}
\value{
Returns list containing:
\item{predictions}{\code{nrow(x_te)} x \code{length(ml)} matrix with the predictions of each method}
\item{weights}{If \code{weights=TRUE}, list of length(ml) with smoother matrices of dimension \code{nrow(xnew)} x \code{nrow(x)}
containing the weights that deliver predictions of each method where each row gives the weight that each training
outcome received in the prediction}
}
\description{
Core function of \code{\link{ensemble}}.
}
\keyword{internal}
