% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{forest_grf_fit}
\alias{forest_grf_fit}
\title{Calculates Random Forest fit using the \code{\link{grf}} package}
\usage{
forest_grf_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to  \code{\link{regression_forest}}}
}
\value{
An object with S3 class "regression_forest"
}
\description{
Calculates Random Forest fit using the \code{\link{grf}} package
}
\keyword{internal}
