% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{handle_weights}
\alias{handle_weights}
\title{Sanitizes potential sample weights}
\usage{
handle_weights(w, n)
}
\arguments{
\item{w}{Vector of weights or null if no weights provided}

\item{n}{number of observations}
}
\value{
Vector of weights
}
\description{
Sanitizes potential sample weights
}
\keyword{internal}
