% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK2_decomposition.R
\name{HK2_result_collector}
\alias{HK2_result_collector}
\title{Tailored collection of parameters produced by \code{\link{HK2_decomposition}}.}
\usage{
HK2_result_collector(
  HK2_decomposition,
  parameter = "adim",
  t_aggregate = c(3, 2),
  x_aggregate = c(3, 2)
)
}
\arguments{
\item{HK2_decomposition}{Object of class \code{\link{HK2_decomposition}}.}

\item{parameter}{Parameter to be collected. Either "adim" or "dim" for the respective decomposition 
or a subset of 1:14. Default "adim".}

\item{t_aggregate}{Integer scalar of vector of length two specifying treatment aggregate to be collected. 
1 is reserved for collection of all treaments. 
If single integer, level is extracted. 
Vector of length two like c(3,2) provides aggregate 2 minus aggregate 1.}

\item{x_aggregate}{Integer scalar of vector of length two specifying covariate aggregate to be collected. 
If single integer, level is extracted. 
1 is reserved for the unconditional version. 
vector of length two like c(3,2) provides aggregate 2 minus aggregate 1.}
}
\description{
Tailored collection of parameters produced by \code{\link{HK2_decomposition}}.
}
