% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_decomposition.R
\name{HK_decomposition}
\alias{HK_decomposition}
\title{Heiler & Knaus decomposition of effect heterogeneity under non-homogeneous treatments}
\usage{
HK_decomposition(
  APO_dml,
  z = NULL,
  intercept = TRUE,
  spline = FALSE,
  subset = NULL,
  ...
)
}
\arguments{
\item{APO_dml}{\code{\link{APO_dml}} object containing the effective treatment specific doubly robust score
and nuisance parameters.}

\item{z}{Vector or matrix containing the low-dimensional heterogeneity variables to be considered. If \code{NULL},
the unconditional decomposition parameters are calculated.}

\item{intercept}{Adds an intercept to the variables in \code{z}. Set \code{FALSE} e.g. to get subgroup effects if
one-hot coded \code{z}.}

\item{spline}{If \code{TRUE}, B-spline regression based on \code{\link{crs}} is conducted.}

\item{subset}{Optional logical vector if decomposition should be run for subsample of original sample.}

\item{...}{Pass option for \code{\link{crs}}.}
}
\value{
\code{HK_decomposition} object with a list of nATE, rATE and Delta list containing each
 \item{score}{The parameter specific Neyman orthogonal score that was used as pseudo-outcome.}
 \item{list}{The \code{\link{lm}} object run to get the least squares coefficients (not with corrected standard errors).}
 \item{vcov}{The adjusted variance-covariance matrix derived in Heiler and Knaus (2021) for rATE and Delta.
 For nATE standard heteroskedasticity robust variance-covariance matrix.}
 \item{results}{A \code{\link{coeftest}} object with the coefficients and the appropriate inference results.}
}
\description{
This function builds on an \code{\link{APO_dml}} object to calculate the decomposition parameters
nATE(Z) = rATE(Z) + Delta(Z) described in Heiler and Knaus (2021).
}
\references{
\itemize{
\item Heiler, P., Knaus, M. C. (2021). Effect or Treatment Heterogeneity?
Policy Evaluation with Aggregated and Disaggregated Treatments
arXiv preprint arXiv:2110.01427. \url{https://arxiv.org/abs/2110.01427}
}
}
