% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneous_effects.R
\name{kr_cate}
\alias{kr_cate}
\title{Estimates non-paramteric CATEs using kernel regression as proposed by Fan et al. (2019)
and Zimmert & Lechner (2019).}
\usage{
kr_cate(delta, z, bw_factor = 0.9)
}
\arguments{
\item{delta}{Vector of doubly robust ATE score. E.g obtained as one column of \code{ATE_dml$delta}
from \code{\link{ATE_dml}}.}

\item{z}{Heterogeneity variable(s) vector, matrix or data.frame.}

\item{bw_factor}{Factor by which cross-validated is multiplied. Default is undersmoothing with
factor 0.9 as recommended by Zimmert & Lechner (2019).}
}
\value{
\code{kr_cate} object:
\item{model}{npqregression object (see\code{\link{npreg}}) of the kernel regression}
\item{fit}{Fitted values of the kernel regression}
\item{bw}{Cross-validated bandwidth (not scaled)}
\item{ate}{Average teratment effect}
}
\description{
Estimates non-paramteric CATEs using kernel regression as proposed by Fan et al. (2019)
and Zimmert & Lechner (2019).
}
\references{
\itemize{
\item Fan, Q., Hsu, Y.-C., Lieli, R. P., & Zhang, Y. (2019). Estimation of conditional averagetreatment effects with high-dimensional data.
arXiv preprint arXiv:1908.02399. \url{http://arxiv.org/abs/1908.02399}
\item Zimmert, M., & Lechner, M. (2019). Nonparametric estimation of causal heterogeneityunder high-dimensional confounding.
arXiv preprint arXiv:1908.02399. \url{http://arxiv.org/abs/1908.08779}
}
}
