% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{lasso_fit}
\alias{lasso_fit}
\title{This function estimates cross-validated lasso regression based on the \code{\link{glmnet}} package}
\usage{
lasso_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to  \code{\link{glmnet}}}
}
\value{
An object with S3 class "glmnet"
}
\description{
This function estimates cross-validated lasso regression based on the \code{\link{glmnet}} package
}
\keyword{internal}
