% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_learner.R
\name{ndr_learner}
\alias{ndr_learner}
\title{DR- and NDR-learner.}
\usage{
ndr_learner(
  y,
  w,
  x,
  ml_w = list(create_method("forest_grf")),
  ml_y = list(create_method("forest_grf")),
  ml_tau = list(create_method("forest_grf")),
  xnew = NULL,
  compare_all = TRUE,
  nfolds = 5,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{y}{Numerical vector containing the outcome variable.}

\item{w}{Treatment vector. Provide as factor to control ordering of the treatments,
otherwise program orders treatments in ascending order or alphabetically.}

\item{x}{Covariate matrix.}

\item{ml_w}{List of methods to be used in ensemble estimation of propensity score.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_y}{List of methods to be used in ensemble estimation of outcome regression.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_tau}{List of methods to be used in ensemble estimation of CATEs.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{xnew}{Covariate matrix of test sample}

\item{compare_all}{Relevant multiple treatments: If FALSE, only comparisons to first treatment considered.}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights (default \code{nfolds=5})}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
IATE objects saved in new subfolder Comparisonij where i is the index of the control and j of the treated group.}

\item{quiet}{If FALSE, ensemble estimators print method that is currently running.}
}
\value{
Returns n x 2 matrix containing DR- and NDR-learner predictions in case of one comparison or
a list n x 2 matrices with the specified comparisons.
}
\description{
This function predicts conditional average treatment effects (CATEs) with \code{\link{ensemble}} using
the DR-learner (Kennedy, 2020) and the normalized DR-learner (Knaus, 2020).
}
\references{
\itemize{
\item Kennedy, E. H. (2020). Optimal doubly robust estimation of heterogeneous causal effects.
arXiv preprint arXiv:2004.14497. \url{http://arxiv.org/abs/2004.14497}
\item Knaus, M. C. (2020). Double machine learning based program evaluation under unconfoundedness.
  arXiv preprint arXiv:2003.03191.\url{http://arxiv.org/abs/2003.03191}
}
}
