% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_learner.R
\name{ndr_oos}
\alias{ndr_oos}
\title{Out-of-sample prediction with DR- and NDR-learner}
\usage{
ndr_oos(
  y,
  w,
  x,
  xnew,
  ml_w = list(create_method("forest_grf")),
  ml_y = list(create_method("forest_grf")),
  ml_tau = list(create_method("forest_grf")),
  cf_mat = NULL,
  compare_all = TRUE,
  nfolds = 5,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{y}{Numerical vector containing the outcome variable.}

\item{w}{Treatment vector. Provide as factor to control ordering of the treatments,
otherwise program orders treatments in ascending order or alphabetically.}

\item{x}{Covariate matrix.}

\item{xnew}{Covariate matrix of test sample.}

\item{ml_w}{List of methods to be used in ensemble estimation of propensity score.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_y}{List of methods to be used in ensemble estimation of outcome regression.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_tau}{List of methods to be used in ensemble estimation of CATEs.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{cf_mat}{Optional logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}). Otherwise created internally.}

\item{compare_all}{Relevant multiple treatments: If FALSE, only comparisons to first treatment considered.}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights (default \code{nfolds=5}).}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
IATE objects saved in new subfolder Comparisonij where i is the index of the control and j of the treated group.}

\item{quiet}{If FALSE, ensemble estimators print method that is currently running.}
}
\value{
\item{cates}{n x 2 matrix containing DR- and NDR-learner predictions in case of a binary treatment or
a list of n x 2 matrices with the specified comparisons.}
         \item{APO}{\code{\link{APO_dml}} object containing the underlying nuisance parameters etc.}
         \item{ATE}{\code{\link{ATE_dml}} object containing the underlying doubly robust score.}
}
\description{
This function produces out-of-sample predictions of conditional average treatment effects (CATEs) using the
DR-learner (Kennedy, 2020) and the normalized DR-learner (Knaus, 2020).
It executes steps 1 to 4 of algorithms 1 and 2 in Knaus (2020) as part of the \code{\link{ndr_learner}}.
}
\references{
\itemize{
\item Kennedy, E. H. (2020). Optimal doubly robust estimation of heterogeneous causal effects.
arXiv preprint arXiv:2004.14497. \url{http://arxiv.org/abs/2004.14497}
\item Knaus, M. C. (2020). Double machine learning based program evaluation under unconfoundedness.
  arXiv preprint arXiv:2003.03191.\url{http://arxiv.org/abs/2003.03191}
}
}
