% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_learner.R
\name{norm_drl}
\alias{norm_drl}
\title{This is a wrapper of the C++ function that normalizes the weights for the NDR-learner}
\usage{
norm_drl(alpha, m_mat, y, w_mat, e_mat)
}
\arguments{
\item{alpha}{Sparse matrix containing the weights for each observation to be predicted}

\item{m_mat}{n x T+1 matrix with fitted outcome values.}

\item{y}{Vector of variable to be predicted.}

\item{w_mat}{Logical matrix of treatment indicators (n x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{e_mat}{n x T+1 matrix with propensity scores.}
}
\value{
NDR-learner CATEs.
}
\description{
This is a wrapper of the C++ function that normalizes the weights for the NDR-learner
}
