% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ols_fit}
\alias{ols_fit}
\title{Calculates OLS fit.}
\usage{
ols_fit(x, y)
}
\arguments{
\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}
}
\value{
Returns OLS coefficients
}
\description{
Calculates OLS fit.
}
\keyword{internal}
