% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{plasso_fit}
\alias{plasso_fit}
\title{This function estimates cross-validated Post-Lasso based on the \code{\link{glmnet}} package}
\usage{
plasso_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to  \code{\link{glmnet}}}
}
\value{
An object with S3 class "plasso"
}
\description{
This function estimates cross-validated Post-Lasso based on the \code{\link{glmnet}} package
}
\keyword{internal}
