% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{plot.APO_dml}
\alias{plot.APO_dml}
\title{\code{plot} method for class \code{\link{APO_dml}}}
\usage{
\method{plot}{APO_dml}(APO_dml, label = NULL, sl = 0.05)
}
\arguments{
\item{APO_dml}{Object of class \code{\link{APO_dml}}.}

\item{label}{Optional vector of treatment labels.}

\item{sl}{Significance level for confidence intervals (default 0.05).}
}
\description{
\code{plot} method for class \code{\link{APO_dml}}
}
