% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_decomposition.R
\name{plot.HK_decomposition}
\alias{plot.HK_decomposition}
\title{\code{plot} coefficients for class \code{\link{HK_decomposition}}}
\usage{
\method{plot}{HK_decomposition}(object, sl = 0.05)
}
\arguments{
\item{object}{Object of class \code{\link{APO_dml}}.}

\item{sl}{Significance level for confidence intervals (default 0.05).}
}
\description{
\code{plot} coefficients for class \code{\link{HK_decomposition}}
}
