% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneous_effects.R
\name{plot.kr_cate}
\alias{plot.kr_cate}
\title{\code{plot} method for class \code{\link{kr_cate}}}
\usage{
\method{plot}{kr_cate}(kr_cate, z_label = NULL, xrange = NULL, yrange = NULL, sl = 0.05)
}
\arguments{
\item{kr_cate}{Object of class \code{\link{kr_cate}}.}

\item{z_label}{Optional label for the x-axis.}

\item{xrange}{Optional range of heterogeneity variable to be plotted on x-axis (e.g. \code{c(-1,1)}).}

\item{yrange}{Optional range of CATE to be plotted on y-axis (e.g. \code{c(-1,1)}).}

\item{sl}{Significance level for confidence intervals (default 0.05).}
}
\description{
\code{plot} method for class \code{\link{kr_cate}}
}
