% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{plot.plasso}
\alias{plot.plasso}
\title{Plot of cross-validation curves.}
\usage{
\method{plot}{plasso}(plasso)
}
\arguments{
\item{plasso}{\code{\link{plasso}} object}
}
\description{
Plot of cross-validation curves.
}
