% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_decomposition.R
\name{predict.HK_decomposition}
\alias{predict.HK_decomposition}
\title{\code{predict} method for class \code{\link{HK_decomposition}}}
\usage{
\method{predict}{HK_decomposition}(object, param = "Delta")
}
\arguments{
\item{object}{\code{\link{HK_decomposition}}}

\item{param}{chose from \code{c("nATE","rATE","Delta")}. Default \code{param="Delta"}}
}
\value{
n times 2 matrix with first column fitted value and second column standard error.
}
\description{
Provides the fitted values of the decomposition parameter of choice including adjusted standard error.
}
