% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{predict.plasso}
\alias{predict.plasso}
\title{Predict after Post-Lasso.}
\usage{
\method{predict}{plasso}(plasso, xnew = NULL, se_rule = 0, weights = FALSE)
}
\arguments{
\item{plasso}{\code{\link{plasso}} object}

\item{xnew}{Matrix of new values for x at which predictions are to be made}

\item{se_rule}{If equal to zero predictions from CV minimum (default). Negative values go in the direction of smaller
models (e.g. se_rule=-1 creates the standard 1SE rule), positive values go in the direction of larger models
(e.g. se_rule=1 creates the standard 1SE+ rule)}

\item{weights}{If TRUE, weights underlying the prediction for xnew calculated}

\item{x}{Covariate matrix that was used for training}

\item{y}{Outcomes that were used for training}
}
\description{
Predict after Post-Lasso.
}
