% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_effects.R
\name{prep_w_mat}
\alias{prep_w_mat}
\title{Creates matrix of binary treament indicators (n x T+1) where each column contains
a binary indicator for one treatment.}
\usage{
prep_w_mat(w)
}
\arguments{
\item{w}{Treatment vector. Provide as factor to control ordering of the treatments,
otherwise program orders treatments in ascending order or alphabetically.}
}
\value{
Logical matrix of treatment indicators (n x T+1).
}
\description{
Creates matrix of binary treament indicators (n x T+1) where each column contains
a binary indicator for one treatment.
}
