% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK2_decomposition.R
\name{print.HK2_decomposition}
\alias{print.HK2_decomposition}
\title{\code{print} method for class \code{\link{HK2_decomposition}} summarizing the setting.}
\usage{
\method{print}{HK2_decomposition}(HK2_decomposition)
}
\arguments{
\item{HK2_decomposition}{Object of class \code{\link{HK2_decomposition}}.}
}
\description{
\code{print} method for class \code{\link{HK2_decomposition}} summarizing the setting.
}
