% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneous_effects.R
\name{spline_cate}
\alias{spline_cate}
\title{Estimates non-paramteric CATEs using spline regression as proposed by
Semenova and Chernozhukov (2020) using the \code{\link{crs}} package.}
\usage{
spline_cate(delta, z, ...)
}
\arguments{
\item{delta}{Vector of doubly robust ATE score. E.g obtained as one column of \code{ATE_dml$delta}
from \code{\link{ATE_dml}}.}

\item{z}{Heterogeneity variable(s) vector, matrix or data.frame}

\item{...}{Pass \code{\link{crs}} options, otherwise default settings are used}
}
\value{
\code{spline_cate} object:
\item{model}{\code{\link{crs}} object of the spline regression}
\item{fit}{Fitted values of the spline regression}
\item{se_fit}{Robust standard errors of fitted values}
\item{ate}{Average teratment effect}
\item{z}{Original heterogeneity variable(s) data.frame}
}
\description{
Estimates non-paramteric CATEs using spline regression as proposed by
Semenova and Chernozhukov (2020) using the \code{\link{crs}} package.
}
\references{
\itemize{
\item Semenova, V., & Chernozhukov, V. (2020). Debiased machine learning of conditional
average treatment effects and other causal functions. The Econometrics Journal,
utaa027, doi: https://doi.org/10.1093/ectj/utaa027
}
}
