% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{summary.APO_dml}
\alias{summary.APO_dml}
\title{\code{summary} method for class \code{\link{APO_dml}}}
\usage{
summary.APO_dml(APO_dml)
}
\arguments{
\item{APO_dml}{Object of class \code{\link{APO_dml}}.}
}
\description{
\code{summary} method for class \code{\link{APO_dml}}
}
