% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{summary.ATE_dml}
\alias{summary.ATE_dml}
\title{\code{summary} method for class \code{\link{ATE_dml}}}
\usage{
summary.ATE_dml(ATE_dml, ...)
}
\arguments{
\item{ATE_dml}{Object of class \code{\link{ATE_dml}}.}

\item{...}{further argumenst passed to \code{printCoefmat}}
}
\description{
\code{summary} method for class \code{\link{ATE_dml}}
}
