% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effects.R
\name{summary.DML_partial_linear}
\alias{summary.DML_partial_linear}
\title{\code{summary} method for class \code{\link{DML_partial_linear}}}
\usage{
\method{summary}{DML_partial_linear}(DML_partial_linear, ...)
}
\arguments{
\item{DML_partial_linear}{Object of class \code{\link{DML_partial_linear}}.}

\item{...}{further arguments passed to \code{printCoefmat}}
}
\description{
\code{summary} method for class \code{\link{DML_partial_linear}}
}
