% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK2_decomposition.R
\name{summary.HK2_decomposition}
\alias{summary.HK2_decomposition}
\title{\code{summary} method for class \code{\link{HK2_decomposition}}}
\usage{
\method{summary}{HK2_decomposition}(
  HK2_decomposition,
  parameter = "adim",
  t_aggregate = c(2, 1),
  x_aggregate = c(3, 2)
)
}
\arguments{
\item{HK2_decomposition}{Object of class \code{\link{HK2_decomposition}}.}

\item{parameter}{Parameter to be collected. Either "adim" or "dim" for the respective decomposition 
or a subset of 1:14. Default "adim".}

\item{t_aggregate}{Integer scalar of vector of length two specifying treatment aggregate to be collected. 
1 is reserved for collection of all treaments. 
If single integer, level is extracted. 
Vector of length two like c(3,2) provides aggregate 2 minus aggregate 1.}

\item{x_aggregate}{Integer scalar of vector of length two specifying covariate aggregate to be collected. 
If single integer, level is extracted. 
1 is reserved for the unconditional version. 
vector of length two like c(3,2) provides aggregate 2 minus aggregate 1.}
}
\value{
Matrix containing point estimates and standard errors of selected parameters.
}
\description{
\code{summary} method for class \code{\link{HK2_decomposition}}
}
