% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_decomposition.R
\name{summary.HK_decomposition}
\alias{summary.HK_decomposition}
\title{\code{summary} method for class \code{\link{HK_decomposition}}}
\usage{
\method{summary}{HK_decomposition}(object)
}
\arguments{
\item{object}{\code{\link{HK_decomposition}}}
}
\description{
Prints the coefficients and inference of the three decomposition parameters.
}
