% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{summary.plasso}
\alias{summary.plasso}
\title{Summary of Post-Lasso model}
\usage{
\method{summary}{plasso}(plasso)
}
\arguments{
\item{plasso}{\code{\link{plasso}} object}
}
\value{
Prints cross-validated MSE and active variables for Lasso and Post-Lasso.
}
\description{
Summary of Post-Lasso model
}
