% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_learner.R
\name{ndr_learner}
\alias{ndr_learner}
\title{DR- and NDR-learner.}
\usage{
ndr_learner(
  y,
  w,
  x,
  ml_w = list(create_method("forest_grf")),
  ml_y = list(create_method("forest_grf")),
  ml_tau = list(create_method("forest_grf")),
  compare_all = TRUE,
  nfolds = 5,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{y}{Numerical vector containing the outcome variable.}

\item{w}{Treatment vector. Provide as factor to control ordering of the treatments,
otherwise program orders treatments in ascending order or alphabetically.}

\item{x}{Covariate matrix.}

\item{ml_w}{List of methods to be used in ensemble estimation of propensity score.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_y}{List of methods to be used in ensemble estimation of outcome regression.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{ml_tau}{List of methods to be used in ensemble estimation of CATEs.
Methods can be created by \code{\link{create_method}}. Default is an untuned honest
\code{\link{regression_forest}}.}

\item{compare_all}{Relevant multiple treatments: If FALSE, only comparisons to first treatment considered.}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights (default \code{nfolds=5}).}

\item{path}{Optional path to save the \code{\link{ensemble}} objects used to predict nuisance parameters and CATEs
for later processing. Outputs of each fold i saved in new subfolder NDR_foldi.}

\item{quiet}{If FALSE, ensemble estimators print method that is currently running.}
}
\value{
\item{cates}{Returns n x 2 matrix containing DR- and NDR-learner predictions in case of a binary treatment or
a number of comparisons x n x 2 array with the predictions.}
         \item{list}{A list of the four \code{\link{ndr_oos}} outputs.}
         \item{cf_mat}{Matrix with k columns of indicators representing the different folds used in estimation.}
}
\description{
This function produces out-of-sample predictions of conditional average treatment effects (CATEs) for
all individuals in the sample using the DR-learner (Kennedy, 2020) and the normalized DR-learner (Knaus, 2020).
The involved predictions are based on \code{\link{ensemble}}. See algorithms 1 and 2 in Knaus (2020) for details.
}
\references{
\itemize{
\item Kennedy, E. H. (2020). Optimal doubly robust estimation of heterogeneous causal effects.
arXiv preprint arXiv:2004.14497. \url{http://arxiv.org/abs/2004.14497}
\item Knaus, M. C. (2020). Double machine learning based program evaluation under unconfoundedness.
  arXiv preprint arXiv:2003.03191.\url{http://arxiv.org/abs/2003.03191}
}
}
